fix: generate embeds before schema/db build and quiet harness knip
Make graphql-schema and db build/typecheck depend on their generate
targets so clean CI checkouts can compile. Move standalone SPA handler
into ready-for-agent and drop unused harness exports.
