+++ packages/db/project.json
@@ -15,6 +15,12 @@
       "outputs": ["{projectRoot}/src/lib/embedded-migrations.gen.ts"],
       "cache": true
     },
+    "build": {
+      "dependsOn": ["generate", "^build"]
+    },
+    "typecheck": {
+      "dependsOn": ["generate", "^build"]
+    },
     "test": {
       "executor": "nx:run-commands",
       "options": {
